requirements.txt
chromadb
fastapi
python-telegram-bot